% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeDirecciones.R
\name{geocodeDirecciones}
\alias{geocodeDirecciones}
\title{Convertí tus direcciones a puntos usando APIs gratuitas y del gobierno argentino}
\usage{
geocodeDirecciones(
  datos = NULL,
  col_direccion = "direccion",
  col_provincia = NULL,
  col_departamento = NULL,
  api = "datosgobar",
  max_returns = 1
)
}
\arguments{
\item{datos}{data.frame data.table tibble REQUERIDO}

\item{col_direccion}{character REQUERIDO. Nombre de columna direccion a geolocalizar y normalizar.}

\item{col_provincia}{character Nombre de columna provincia a geolocalizar y normalizar. Es opcional en caso de querer precisar mayor información para el geocoding}

\item{col_departamento}{character Nombre de columna departamento a geolocalizar y normalizar. Es opcional en caso de querer precisar mayor información para el geocoding}

\item{api}{text Filtrar por nombre o ID de departamento.}

\item{max_returns}{numeric. Cantidad de devoluciones máximas de geolocalización por dirección. Se encuentra puesto en 1}
}
\value{
Un objeto data.frame con información sobre la ubicación espacial de direcciones de Argentina
}
\description{
Convertí tus direcciones a puntos usando APIs gratuitas y del gobierno argentino
}
\examples{
\dontrun{
df <- datos(direccion=c("Figueroa Alcorta 7597","Pedro Goyena 330"),
            provincia=c("CABA","CABA"))
geocodeDirecciones(datos=df,col_direccion="direccion",col_provincia="provincia")
}
}
