% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeGoogle.R
\name{geocodeGoogle}
\alias{geocodeGoogle}
\title{Convertí las direcciones a coordenadas usando la API de Google}
\usage{
geocodeGoogle(
  datos = NULL,
  col_key = NULL,
  col_id = NULL,
  cols_query = NULL,
  col_region = NULL
)
}
\arguments{
\item{datos}{obligatorio: data.frame o data.table con los datos a geolocalizar}

\item{col_key}{obligatorio: columna con la key para usar en cada una de las filas. Útil para superar las restricciones de Google}

\item{col_id}{obligatorio:}

\item{cols_query}{obligatorio: Columnas que van a agregarse a la consulta de google}

\item{col_region}{opcional: Es posible agregar una columna con la región para forzar resultados en esa región (por ejemplo "AR")}
}
\value{
Un objeto data.frame con información sobre la ubicación espacial de direcciones de Argentina
}
\description{
Convertí las direcciones a coordenadas usando la API de Google
}
\examples{
\dontrun{
df <- data.frame(direccion=c("Figueroa Alcorta 7597","Pedro Goyena 330"),
            provincia=c("CABA","CABA"))
df$id <- 1:nrow(df)
df$key <-"alkfnaf901ehfassafqZ"
geocodeGoogle(datos=df,col_id="id",col_key="key", cols_query=c("direccion","provincia"),region=NULL)

# Restringiendo resultados a Argentina
df$reg <- "AR"
geocodeGoogle(datos=df,col_id="id",col_key="key", cols_query=c("direccion","provincia"),region="reg")

# Hay que cambiar el valor de la key por uno real para que funcione
}
}
