% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeDirecciones.R
\name{geocodeDirecciones}
\alias{geocodeDirecciones}
\title{Convertí tus direcciones a puntos usando APIs gratuitas y del gobierno argentino}
\usage{
geocodeDirecciones(
  datos = NULL,
  col_id = NULL,
  col_direccion = "direccion",
  col_provincia = NULL,
  col_departamento = NULL,
  api = "datosgobar",
  max_returns = 1
)
}
\arguments{
\item{datos}{OBLIGATORIO: Es un data.frame/data.table/tibble que contiene información sobre las direcciones a geolocalizar  (data.frame/data.table/tibble)}

\item{col_id}{OBLIGATORIO: Nombre de la columna que contiene un id único por dirección. Aunque la función no hace un chequeo de que sea único, es necesario que así lo sea para poder unir la devolución de puntos, que puede ser múltiple. (Vector numérico o character)}

\item{col_direccion}{OBLIGATORIO: Nombre de la columna que contiene la dirección a geolocalizar y normalizar. Por ejemplo, que contenga valores como "Figueroa Alcorta 7597".  (Vector character)}

\item{col_provincia}{Nombre de la columna que contiene la provincia de la direccion geolocalizar y normalizar. Por ejemplo, que contenga valores como "Ciudad de Buenos Aires". (Vector character)}

\item{col_departamento}{Nombre de la columna que contiene el departamento de la direccion geolocalizar y normalizar. Por ejemplo, que contenga valores como "Comuna 11" o "Berazategui". (Vector character)}

\item{api}{Texto que dice que API usar. "datosgobar" es la API del gobierno nacional (https://datosgobar.github.io/georef-ar-api/). "gcba") es la API del gobierno de la Ciudad de Buenos Aires (pendiente de implementación)}

\item{max_returns}{Cantidad de devoluciones máximas de geolocalización por dirección.}
}
\value{
Un objeto data.frame con información sobre la ubicación espacial de direcciones de Argentina
}
\description{
Convertí tus direcciones a puntos usando APIs gratuitas y del gobierno argentino
}
\examples{
\dontrun{
df <- data.frame(direccion=c("Figueroa Alcorta 7597","Pedro Goyena 330"),
            provincia=c("CABA","CABA"))
geocodeDirecciones(datos=df,col_direccion="direccion",col_provincia="provincia")
}
}
